--- server/src/schemaTypes/mutation.ts/channels/closeChannel.ts
@@ -1,15 +1,0 @@
-import { GraphQLObjectType, GraphQLString } from "graphql";
-
-export const CloseChannelType = new GraphQLObjectType({
-  name: "closeChannelType",
-  fields: () => {
-    return {
-      transactionId: {
-        type: GraphQLString
-      },
-      transactionOutputIndex: {
-        type: GraphQLString
-      }
-    };
-  }
-});
