--- server/src/schemaTypes/mutation.ts/channels/openChannel.ts
@@ -1,15 +1,0 @@
-import { GraphQLObjectType, GraphQLString } from "graphql";
-
-export const OpenChannelType = new GraphQLObjectType({
-  name: "openChannelType",
-  fields: () => {
-    return {
-      transactionId: {
-        type: GraphQLString
-      },
-      transactionOutputIndex: {
-        type: GraphQLString
-      }
-    };
-  }
-});
