--- server/src/schemaTypes/mutation.ts/invoice/createInvoice.ts
@@ -1,17 +1,0 @@
-import { GraphQLObjectType, GraphQLString, GraphQLInt } from "graphql";
-import { GraphQLDateTime } from "graphql-iso-date";
-
-export const InvoiceType = new GraphQLObjectType({
-  name: "invoiceType",
-  fields: () => {
-    return {
-      chainAddress: { type: GraphQLString },
-      createdAt: { type: GraphQLDateTime },
-      description: { type: GraphQLString },
-      id: { type: GraphQLString },
-      request: { type: GraphQLString },
-      secret: { type: GraphQLString },
-      tokens: { type: GraphQLInt }
-    };
-  }
-});
