+++ server/src/schemaTypes/mutation/channels/closeChannel.ts
@@ -1,0 +1,15 @@
+import { GraphQLObjectType, GraphQLString } from 'graphql';
+
+export const CloseChannelType = new GraphQLObjectType({
+    name: 'closeChannelType',
+    fields: () => {
+        return {
+            transactionId: {
+                type: GraphQLString,
+            },
+            transactionOutputIndex: {
+                type: GraphQLString,
+            },
+        };
+    },
+});
