+++ server/src/schemaTypes/mutation/channels/openChannel.ts
@@ -1,0 +1,15 @@
+import { GraphQLObjectType, GraphQLString } from 'graphql';
+
+export const OpenChannelType = new GraphQLObjectType({
+    name: 'openChannelType',
+    fields: () => {
+        return {
+            transactionId: {
+                type: GraphQLString,
+            },
+            transactionOutputIndex: {
+                type: GraphQLString,
+            },
+        };
+    },
+});
