+++ server/src/schemas/mutations/invoices/createInvoice.ts
@@ -2,7 +2,7 @@ import { createInvoice as createInvoiceRequest } from 'ln-service';
 import { logger } from '../../../helpers/logger';
 import { requestLimiter } from '../../../helpers/rateLimiter';
 import { GraphQLNonNull, GraphQLInt } from 'graphql';
-import { InvoiceType } from '../../../schemaTypes/mutation.ts/invoice/createInvoice';
+import { InvoiceType } from '../../../schemaTypes/mutation/invoice/createInvoice';
 import { getErrorMsg, getAuthLnd } from '../../../helpers/helpers';
 import { defaultParams } from '../../../helpers/defaultProps';
