+++ server/src/schemas/mutations/invoices/parsePayment.ts
@@ -2,7 +2,7 @@ import { parsePaymentRequest } from 'ln-service';
 import { logger } from '../../../helpers/logger';
 import { requestLimiter } from '../../../helpers/rateLimiter';
 import { GraphQLString, GraphQLNonNull } from 'graphql';
-import { ParsePaymentType } from '../../../schemaTypes/mutation.ts/invoice/parsePayment';
+import { ParsePaymentType } from '../../../schemaTypes/mutation/invoice/parsePayment';
 import { getErrorMsg, getAuthLnd } from '../../../helpers/helpers';
 import { defaultParams } from '../../../helpers/defaultProps';
 
@@ -47,7 +47,7 @@ export const parsePayment = {
                 request: params.request,
             });
 
-            const routes = request.routes.map(route => {
+            const routes = request.routes.map((route) => {
                 return {
                     mTokenFee: route.base_fee_mtokens,
                     channel: route.channel,
