+++ server/src/schemas/mutations/invoices/pay.ts
@@ -2,7 +2,7 @@ import { pay as payRequest } from 'ln-service';
 import { logger } from '../../../helpers/logger';
 import { requestLimiter } from '../../../helpers/rateLimiter';
 import { GraphQLString, GraphQLNonNull } from 'graphql';
-import { PayType } from '../../../schemaTypes/mutation.ts/invoice/pay';
+import { PayType } from '../../../schemaTypes/mutation/invoice/pay';
 import { getErrorMsg, getAuthLnd } from '../../../helpers/helpers';
 import { defaultParams } from '../../../helpers/defaultProps';
 
@@ -44,7 +44,7 @@ export const pay = {
                 request: params.request,
             });
 
-            const hops = payment.hops.map(hop => {
+            const hops = payment.hops.map((hop) => {
                 return {
                     channel: hop.channel,
                     channelCapacity: hop.channel_capacity,
